Add hero animation to book images
This commit adds hero animation to book images in the `BookDetailsView` and `ImageList` widgets. The `ImageContainer` widget now takes a `heroTag` parameter to specify the tag for the hero animation. This improves the user experience by providing a smooth transition between screens when the user taps on a book image. Additionally, the `pubspec.lock` file has been updated to include the latest versions of some dependencies.